# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AARRR Dashboard Builder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface Stage {
  stageKey: string;
  customLabel: string;
  order: number;
  metricValue?: number | null;
  conversionRate?: number | null;
  dropOffRate?: number | null;
  note?: string | null;
}

export interface Project {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  isHidden: boolean;
  createdAt: string;
  dashboardCount: number;
}

export interface Dashboard {
  id: number;
  projectId: number;
  title: string;
  slug: string;
  serviceName?: string | null;
  createdByToken: string;
  createdAt: string;
  updatedAt: string;
  isHidden: boolean;
  stages: Stage[];
}

export interface CreateProjectInput {
  name: string;
  slug?: string | null;
  description?: string | null;
  isHidden?: boolean;
}

export interface CreateDashboardInput {
  title: string;
  slug?: string | null;
  serviceName?: string | null;
  createdByToken: string;
  stages: Stage[];
}

export interface UpdateDashboardInput {
  title: string;
  serviceName?: string | null;
  ownerToken: string;
  stages: Stage[];
}