# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AARRR Dashboard Builder API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Project operations
  - name: dashboards
    description: Dashboard operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all visible projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /projects/{projectSlug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by slug
      parameters:
        - name: projectSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project (admin only)
      parameters:
        - name: projectSlug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectInput"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (admin only)
      parameters:
        - name: projectSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /projects/{projectSlug}/dashboards:
    get:
      operationId: listDashboards
      tags: [dashboards]
      summary: List dashboards for a project
      parameters:
        - name: projectSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of dashboards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dashboard"
    post:
      operationId: createDashboard
      tags: [dashboards]
      summary: Create a dashboard under a project
      parameters:
        - name: projectSlug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDashboardInput"
      responses:
        "201":
          description: Created dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"

  /projects/{projectSlug}/dashboards/{dashboardSlug}:
    get:
      operationId: getDashboard
      tags: [dashboards]
      summary: Get a dashboard by slug
      parameters:
        - name: projectSlug
          in: path
          required: true
          schema:
            type: string
        - name: dashboardSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateDashboard
      tags: [dashboards]
      summary: Update a dashboard
      parameters:
        - name: projectSlug
          in: path
          required: true
          schema:
            type: string
        - name: dashboardSlug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDashboardInput"
      responses:
        "200":
          description: Updated dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDashboard
      tags: [dashboards]
      summary: Delete a dashboard (admin only)
      parameters:
        - name: projectSlug
          in: path
          required: true
          schema:
            type: string
        - name: dashboardSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    Stage:
      type: object
      properties:
        stageKey:
          type: string
        customLabel:
          type: string
        order:
          type: integer
        metricValue:
          type: number
          nullable: true
        conversionRate:
          type: number
          nullable: true
        dropOffRate:
          type: number
          nullable: true
        note:
          type: string
          nullable: true
      required:
        - stageKey
        - customLabel
        - order

    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        isHidden:
          type: boolean
        createdAt:
          type: string
        dashboardCount:
          type: integer
      required:
        - id
        - name
        - slug
        - isHidden
        - createdAt
        - dashboardCount

    Dashboard:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        slug:
          type: string
        serviceName:
          type: string
          nullable: true
        createdByToken:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        isHidden:
          type: boolean
        stages:
          type: array
          items:
            $ref: "#/components/schemas/Stage"
      required:
        - id
        - projectId
        - title
        - slug
        - createdByToken
        - createdAt
        - updatedAt
        - isHidden
        - stages

    CreateProjectInput:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        isHidden:
          type: boolean
      required:
        - name

    CreateDashboardInput:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
          nullable: true
        serviceName:
          type: string
          nullable: true
        createdByToken:
          type: string
        stages:
          type: array
          items:
            $ref: "#/components/schemas/Stage"
      required:
        - title
        - createdByToken
        - stages

    UpdateDashboardInput:
      type: object
      properties:
        title:
          type: string
        serviceName:
          type: string
          nullable: true
        ownerToken:
          type: string
        stages:
          type: array
          items:
            $ref: "#/components/schemas/Stage"
      required:
        - title
        - ownerToken
        - stages