# lib/api-zod/src/generated/types/createDashboardInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AARRR Dashboard Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { Stage } from "./stage";

export interface CreateDashboardInput {
  title: string;
  slug?: string | null;
  serviceName?: string | null;
  createdByToken: string;
  stages: Stage[];
}