# lib/api-zod/src/generated/types/createProjectInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AARRR Dashboard Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProjectInput {
  name: string;
  slug?: string | null;
  description?: string | null;
  isHidden?: boolean;
}