# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AARRR Dashboard Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { Stage } from "./stage";

export interface Dashboard {
  id: number;
  projectId: number;
  title: string;
  slug: string;
  serviceName?: string | null;
  createdByToken: string;
  createdAt: string;
  updatedAt: string;
  isHidden: boolean;
  stages: Stage[];
}