# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AARRR Dashboard Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  isHidden: boolean;
  createdAt: string;
  dashboardCount: number;
}