# lib/api-zod/src/generated/types/stage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AARRR Dashboard Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface Stage {
  stageKey: string;
  customLabel: string;
  order: number;
  metricValue?: number | null;
  conversionRate?: number | null;
  dropOffRate?: number | null;
  note?: string | null;
}