# lib/api-zod/src/generated/types/updateDashboardInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AARRR Dashboard Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { Stage } from "./stage";

export interface UpdateDashboardInput {
  title: string;
  serviceName?: string | null;
  ownerToken: string;
  stages: Stage[];
}